Clean up client and extra deps (#174)
* Add in the actual optional dependencies for client
* Remove unnecessary info from the README
* Move underlying module from `concrete_core` -> `concrete`
* Added a simple chat command `uv run python -m concrete chat "hello"`
* Deleted KnowledgeGraphTool


Tested
* db integration with cli `prompt`

---------
