+++ src/concrete-async/concrete_async/__init__.py
@@ -1,15 +1,17 @@
 from collections.abc import Callable
+from typing import Any, cast
 
-import concrete_core
 from celery.result import AsyncResult
 
-from .models import KombuMixin, Operation
-from .tasks import abstract_operation
+import concrete
+from concrete.clients import CLIClient, model_to_schema
+from concrete.models import KombuMixin, Message, Operation
+from concrete.tasks import abstract_operation
 
 
 def _delay_factory(string_func: Callable[..., str]) -> Callable[..., AsyncResult]:
     def _delay(
-        self: concrete_core.abstract.AbstractOperator,
+        self: concrete.abstract.AbstractOperator,
         *args,
         options: dict,
         **kwargs,
@@ -19,7 +21,7 @@ def _delay_factory(string_func: Callable[..., str]) -> Callable[..., AsyncResult
                 {"role": "system", "content": options["instructions"]},
                 {"role": "user", "content": string_func(self, *args, **kwargs)},
             ],
-            "message_format": concrete_core.clients.OpenAIClient.model_to_schema(options["response_format"]),
+            "message_format": model_to_schema(options["response_format"]),
         }
         operation = Operation(
             client_name=self.llm_client,
@@ -29,7 +31,7 @@ def _delay_factory(string_func: Callable[..., str]) -> Callable[..., AsyncResult
 
         clients = {}
         for name, client in self.clients.items():
-            client_model = concrete_core.models.clients.OpenAIClientModel(
+            client_model = concrete.models.clients.OpenAIClientModel(
                 model=client.model,
                 temperature=client.default_temperature,
             )
@@ -42,26 +44,26 @@ def _delay_factory(string_func: Callable[..., str]) -> Callable[..., AsyncResult
     return _delay
 
 
-for operator_name, operator in concrete_core.abstract.AbstractOperatorMetaclass.OperatorRegistry.items():
+for operator_name, operator in concrete.abstract.AbstractOperatorMetaclass.OperatorRegistry.items():
     for attr, method in operator.__dict__.items():
         if attr.startswith("__") or attr in {"_qna", "qna"} or not callable(method):
             continue
         setattr(method, "_delay", _delay_factory(method))
 
 
-for message_name, message in concrete_core.models.messages.MESSAGE_REGISTRY.items():
+for message_name, message in concrete.models.messages.MESSAGE_REGISTRY.items():
     new_class = type(
         message.__name__,
-        (KombuMixin, message),
+        (cast(type, KombuMixin), cast(Any, message)),
         {'__module__': message.__module__},
     )
-    setattr(concrete_core.models.messages, message.__name__, new_class)
-    concrete_core.models.messages.MESSAGE_REGISTRY[message_name] = new_class
+    setattr(concrete.models.messages, message.__name__, new_class)
+    concrete.models.messages.MESSAGE_REGISTRY[message_name] = cast(Message, new_class)
 
-original_model = concrete_core.models.clients.OpenAIClientModel
+original_model = concrete.models.clients.OpenAIClientModel
 
 setattr(
-    concrete_core.models.clients,
+    concrete.models.clients,
     original_model.__name__,
     type(
         original_model.__name__,
@@ -69,3 +71,5 @@ setattr(
         {'__module__': original_model.__module__},
     ),
 )
+
+CLIClient.emit("concrete-async initialized")
