+++ src/concrete-async/concrete_async/tasks.py
@@ -1,6 +1,6 @@
 from typing import Any, Callable, cast
 
-import concrete_core
+import concrete
 
 from .celery import app
 from .models import ConcreteChatCompletion, KombuMixin, Operation
@@ -11,7 +11,7 @@ def abstract_operation(operation: Operation, clients: dict[str, KombuMixin]) -> 
     """
     An operation that's able to execute arbitrary methods on operators/agents
     """
-    client = concrete_core.clients.OpenAIClient(**clients[operation.client_name].model_dump())
+    client = concrete.clients.OpenAIClient(**clients[operation.client_name].model_dump())
     func: Callable[..., Any] = getattr(client, operation.function_name)
 
     res = func(**operation.arg_dict).model_dump()
