+++ src/concrete-core/pyproject.toml
@@ -6,16 +6,29 @@ readme = "README.md"
 requires-python = ">=3.11.9, <4.0.0"
 
 dependencies = [
-    "python-dotenv>=1.0.1",
-    "pydantic>=2.9.2",
-    "requests>=2.32.3",
+    "python-dotenv>=1.0.0",
+    "pydantic>=2.0.0",
 ]
 
 [project.optional-dependencies]
-openai = [
-    "openai>=1.40.0",
+openai = ["openai>=1.40.0"]
+requests = ["requests"]
+auth = [
+    "pyjwt",
+    "cryptography",
+    "itsdangerous",  # session middleware
+    "fastapi"
 ]
+google-auth = [
+    "google-auth",
+    "google-auth-oauthlib",
+    "google-api-python-client",
+]
+
 
 [build-system]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["concrete"]
