+++ src/concrete-db/concrete_db/orm/setup.py
@@ -1,12 +1,13 @@
 import os
 from contextlib import contextmanager
 
-from concrete_core.clients import CLIClient
 from dotenv import load_dotenv
 from sqlalchemy import URL
 from sqlalchemy.orm import sessionmaker
 from sqlmodel import Session as SQLModelSession
 from sqlmodel import create_engine
+
+from concrete.clients import CLIClient
 
 load_dotenv(override=True)
 if (
